pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/masinag/allsat-cnf"
Documentation = "https://github.com/masinag/allsat-cnf"
Repository = "https://github.com/masinag/allsat-cnf"

[project]
authors = [
    { name = "Gabriele Masina" }
]
name = "allsat_cnf"
version = "1.0"
description = "CNF encodings for AllSAT and AllSMT"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["allsat", "allsmt", "cnf", "sat", "smt"]
classifiers = [
    "Programming Language :: Python :: 3",
]

dependencies = [
    "PySMT >= 0.9.6.dev53", # mathsat version >= 5.6.7
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-benchmark",
    "pytest-runner",
]

benchmark = [
    "circuitgraph >= 0.2.0",
    "funcy >= 1.17",
    "GitPython",
    "matplotlib",
    "pandas",
    "psutil>=5.9.4",
    "py-aiger >= 6.2.0",
    "tqdm",
    "wmipa @ git+https://github.com/unitn-sml/wmi-pa#egg=wmipa",
    "wmibench @ git+https://github.com/unitn-sml/wmi-benchmarks#egg=wmibench",
]




